requirements.txt
asgiref==3.8.1
Django==5.1.3
django-socio-grpc==0.23.1
djangorestframework==3.15.2
Faker==33.1.0
grpcio==1.68.0
grpcio-health-checking==1.68.0
grpcio-tools==1.68.0
gunicorn==23.0.0
lark==1.2.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
protobuf==5.29.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
